Add config file. Add automatic download of application icon

+++ firefox-ssb.py
@@ -1,5 +1,5 @@
 from pathlib import Path
-from configparser import RawConfigParser
+from configparser import ConfigParser, RawConfigParser
 from mozprofile import profile
 import os
 import sys
@@ -7,86 +7,118 @@ import logging
 import argparse
 import json
 import shutil
+import requests
+import coloredlogs
 
+logger = logging.getLogger()
 
 class SSBManager():
-    __launchers_path = None
-    __profiles_loc = None
+    __CONFIG_PATH = './config.ini'
+    __config = None
 
-    APP_JSON_PATH = './data/app.json'
-    LAUNCHERS_PATH = '.local/share/applications'
-    PROFILES_PATH = '.firefox-ssb'
+    __app_json_location = None
+    __launchers_location = None
+    __profiles_location = None
+    __icons_location = None
 
     __app_data = None
 
     def __init__(self):
-        logging.debug("Check the existence of apps.json file")
+        logger.debug("Opening config.ini")
+        config_obj = ConfigParser()
+        config_obj.read(self.__CONFIG_PATH)
+        config_obj.sections()
+        self.__config = config_obj['CONFIG']
+        
+        logger.debug("Loading relevant paths")
+        self.__app_json_location = Path(self.__config['AppJsonLocation'])
 
-        if Path(self.APP_JSON_PATH).exists():
-            logging.debug("app.json file already exists")
-            with open(self.APP_JSON_PATH) as json_file:
+        self.__launchers_location = Path('{0!s}/{1!s}'.format(str(Path.home()), self.__config['LaunchersLocation']))
+        logger.debug("Desktop files will be saved in %s",self.__launchers_location)
+        self.__profiles_location = Path('{0!s}/{1!s}'.format(str(Path.home()), self.__config['ProfilesLocation']))
+        logger.debug("Profiles will be saved in %s",self.__launchers_location)
+        self.__icons_location = Path('{0!s}/{1!s}'.format(str(Path.home()), self.__config['IconsLocation']))
+        logger.debug("Icons will be saved in %s",self.__icons_location)
+
+        logger.debug("Check the existence of apps.json file")
+        if Path(self.__app_json_location).exists():
+            logger.debug("app.json file already exists")
+            with open(self.__app_json_location) as json_file:
                 self.__app_data = json.load(json_file)
         else:
-            logging.debug("app.json file doesn't exists. We will create it...")
+            logger.debug("app.json file doesn't exist. We will create it...")
             self.__app_data = dict()
 
-        self.__launchers_path = Path("{0!s}/{1!s}".format(str(Path.home()), self.LAUNCHERS_PATH))
-        logging.debug("Path of desktop files: %s",self.__launchers_path)
-        logging.info("Profiles will be saved in %s", self.PROFILES_PATH)
-        self.__profiles_loc = Path(Path.home()) / self.PROFILES_PATH
-        if self.__profiles_loc.exists():
-            logging.info(".firefox-ssb directory already exists")
+        logger.debug("Check if firefox profiles directory already exists")
+        if self.__profiles_location.exists():
+            logger.debug(".firefox-ssb directory already exists")
         else:
-            logging.info(".firefox-ssb directory is missing. Creating new one...")
+            print(".firefox-ssb directory is missing. Creating new one...")
             try:
-                os.mkdir(str(self.__profiles_loc))
+                os.mkdir(str(self.__profiles_location))
             except:
-                logging.error("An error occurred during directory creation.")
+                logger.error("An error occurred during directory creation.")
         
     
     def install_app(self, app_name, url, icon_path=None):
-        logging.debug("Check if the application is already installed")
+        logger.debug("Check if the application is already installed")
         if app_name in self.__app_data.keys():
-            logging.error("%s is already installed. Aborting.", app_name)
+            logger.error("%s is already installed. Aborting.", app_name)
             sys.exit()
         
         launcher_file_name = "{!s}.desktop".format(app_name)
         if Path(launcher_file_name).exists():
-            logging.error("The desktop entry already exists. Maybe the web application is already installed. Arresting...")
+            logger.error("The desktop entry already exists. Maybe the web application is already installed. Arresting...")
             sys.exit()
 
-        logging.info("Creating new Firefox profile")
-        profile_path = self.__profiles_loc / app_name
+        print("Creating new Firefox profile")
+        profile_path = self.__profiles_location / app_name
         new_profile = profile.FirefoxProfile(profile=profile_path, restore=False)
 
-        logging.info("New profile created in: %s",new_profile.profile)
-        logging.debug("Setting default profile preferences")
-        logging.debug("Enabling ssb")
+        print("New profile created in: {!s}".format(new_profile.profile))
+        logger.debug("Setting default profile preferences")
+        logger.debug("Enabling ssb")
 
         new_profile.set_persistent_preferences({'browser.ssb.enabled':True})
-        logging.debug("Disabling resume_from_crash")
+        logger.debug("Disabling resume_from_crash")
         new_profile.set_persistent_preferences({'browser.sessionstore.resume_from_crash':False})
-        logging.debug("Disabling disk caching")
+        logger.debug("Disabling disk caching")
         new_profile.set_persistent_preferences({'browser.cache.disk.enable':False})
 
-        logging.debug("Saving profile path")
+        logger.debug("Saving profile path")
         self.__app_data[app_name] = dict()
         self.__app_data[app_name]['profile'] = str(profile_path)
        
         config_obj = RawConfigParser()
         config_obj.optionxform = str
 
-        logging.info("Creating %s file in %s", launcher_file_name, self.__launchers_path)
-        launcher_path = str(self.__launchers_path / launcher_file_name)
+        print("Creating {0!s} file in {1!s}".format(launcher_file_name, self.__launchers_location))
+        launcher_path = str(self.__launchers_location / launcher_file_name)
         config_obj.read(launcher_path)
-        logging.info("Creating Desktop Entry")
+        print("Creating Desktop Entry")
         section_name = 'Desktop Entry'
         
         config_obj.add_section(section_name)
         config_obj[section_name]['Comment'] = ''
         config_obj[section_name]['Exec'] = 'firefox --no-remote --profile {0!s} --class {1!s} --ssb {2!s}'.format(new_profile.profile, app_name, url)
 
-        config_obj[section_name]['Icon'] = icon_path
+        if icon_path is None:
+            logger.debug("Trying to download the favicon")
+            r = requests.get('{!s}/favicon.ico'.format(url))
+            icon = r.content
+            logger.debug("Downloaded icon: %s",icon)
+            if icon is not None:
+                icon_file_name = "{!s}.ico".format(app_name.lower())
+                logger.debug("The name of the icon file will be %s.ico", icon_file_name)
+                icon_path = self.__icons_location / icon_file_name
+                logger.debug("Saving the icon in %s", icon_path)
+                with open(icon_path,'wb') as f:
+                    f.write(icon)
+            else:
+                logger.warning("No icon found. Default icon will be used.")
+                icon = ''
+
+        config_obj[section_name]['Icon'] = str(icon_path)
         config_obj[section_name]['Name'] = app_name
         config_obj[section_name]['NoDisplay'] = 'false'
         config_obj[section_name]['Path[$e]'] = ''
@@ -95,40 +127,46 @@ class SSBManager():
         config_obj[section_name]['TerminalOptions'] = ''
         config_obj[section_name]['Type'] = 'Application'
         config_obj[section_name]['StartupWMClass'] = app_name
-        logging.info("Saving .desktop file")
+        print("Saving .desktop file")
         
 
         with open(launcher_path, 'w') as configfile:    # save
             config_obj.write(configfile,space_around_delimiters=False)
         self.__app_data[app_name]['desktop'] = str(launcher_path)
 
-        logging.debug("Saving additional information")
-        logging.debug("Saving URL")
+        logger.debug("Saving additional information")
+        logger.debug("Saving URL")
         self.__app_data[app_name]['URL'] = url
-        logging.debug("Saving icon path")
-        self.__app_data[app_name]['icon'] = icon_path
+        logger.debug("Saving icon path")
+        self.__app_data[app_name]['icon'] = str(icon_path)
 
-        logging.info("Saving application data")
+        print("Saving application data")
         
-        with open(self.APP_JSON_PATH, 'w') as json_file:
+        with open(self.__app_json_location, 'w') as json_file:
             json.dump(self.__app_data, json_file)
 
-        logging.info("%s application has been installed correctly!", app_name)
+        print("{!s} application has been installed correctly!".format(app_name))
 
 
     def uninstall_app(self, app_name):
-        logging.debug("Check if application exists")
+        logger.debug("Check if application exists")
         if app_name in self.__app_data.keys():
-            logging.info("Removing %s file", self.__app_data[app_name]['desktop'])
+            print('{!s} will be removed. Proceed? [y/N]'.format(app_name), end=' ')
+            if input() != 'y':
+                print('Aborting...')
+                sys.exit()
+            print("Removing {!s} file".format(self.__app_data[app_name]['desktop']))
             os.remove(self.__app_data[app_name]['desktop'])
-            logging.info("Removing firefox profile")
+            print("Removing firefox profile")
             shutil.rmtree(self.__app_data[app_name]['profile'], ignore_errors=True)
-            logging.info("Remove application entry")
+            print("Remove application entry")
             del self.__app_data[app_name]
-            logging.info("Saving changes")
-            with open(self.APP_JSON_PATH, 'w') as json_file:
+            print("Saving changes")
+            with open(self.__app_json_location, 'w') as json_file:
                 json.dump(self.__app_data, json_file)
-        logging.error("%s is not installed.")
+            print("Successfully uninstalled {!s}".format(app_name))
+        else:
+            logger.error("%s is not installed.", app_name)
 
     def list_apps(self):
         print("-"*85)
@@ -140,46 +178,50 @@ class SSBManager():
 
 
 if __name__ == '__main__':
-    #argv = sys.argv[1:]
-    
+    console = logging.StreamHandler()
+    logger.addHandler(console)
+    coloredlogs.install(level='INFO', logger=logger)
+
     app_name = None
     url = None
     icon_path = None
 
     parser=argparse.ArgumentParser(description='''Firefox-SSB''', epilog=""".""")
-    parser.add_argument('action', choices=['install','uninstall','list'], help='')
-    #parser.add_argument('uninstall', action="store_true", help='uninstall a web application')
-    #parser.add_argument('list', action="store_true", help='get the list of the installed applications')
+    parser.add_argument('action', choices=['install','uninstall','list'], help='action to perform')
     parser.add_argument('-n','--name', type=str, nargs=1, help='name to assign to the web application')
     parser.add_argument('-u','--url', type=str, nargs=1, help='URL pointing to the web application to install')
-    parser.add_argument('-i','--icon', type=str, nargs='?', default='./icons/webapp.ico',help='path of the icon to associate with the web application')
+    parser.add_argument('-i','--icon', type=str, nargs='?',help='path of the icon to associate with the web application')
     parser.add_argument('-d','--debug',  action="store_true", help='Enable debug mode')
     args = parser.parse_args()
     
     action = args.action
     if action in ('install','uninstall'):
         if args.name is None:
-            logging.error("You must specify a name for the application")
+            logger.error("You must specify a name for the application")
             parser.print_help()
             sys.exit()
         app_name = args.name.pop()
 
     if action in ('install'):
         if args.url is None:
-            logging.error("You must specify a valid URL to the web application")
+            logger.error("You must specify a valid URL to the web application")
             parser.print_help()
             sys.exit()
         url = args.url.pop()
     
     if action in ('install'):
-        icon_path = args.icon
+        if args.icon is None:
+            print("You haven't provided a valid icon. We will try to find it")
+            icon_path = None
+        else:
+            icon_path = args.icon
 
-    debug_level = logging.INFO
     if args.debug:
         debug_level = logging.DEBUG
-    logging.basicConfig(level=debug_level, format='[%(levelname)-s] - %(message)s')
+        logger.setLevel(debug_level)
 
     manager = SSBManager()
+    
     if args.action == 'install':
         manager.install_app(app_name, url, icon_path)
